Agregando endpoints

+++ src/main/java/ucb/edu/bo/project/repository/BookRepository.java
@@ -16,10 +16,10 @@ public interface BookRepository extends PagingAndSortingRepository<Book, Integer
     List<Book> getBooksJPA();
 
     @Query(value = "SELECT b FROM Book b WHERE b.bookId = :id")
-    Book getBookByIdJPA(Long id);
+    Book getBookByIdJPA(Integer id);
 
     @Query(value = "SELECT * FROM book b WHERE b.book_id = :id", nativeQuery = true)
-    Book getBookById(Long id);
+    Book getBookById(Integer id);
